refactor array list
