+++ src/DataStructures/CountTrait.php
@@ -1,0 +1,18 @@
+<?php
+
+declare(strict_types=1);
+
+namespace ArangoDB\DataStructures;
+
+trait CountTrait
+{
+    /**
+     * Count elements of an object
+     *
+     * @return int
+     */
+    public function count(): int
+    {
+        return count($this->content);
+    }
+}
